==> jupyter_notebooks/Modelo 16/02_gestion_nulos.ipynb <==
# Tratamiento de datos
# -----------------------------------------------------------------------
import numpy as np
import pandas as pd
pd.set_option('display.float_format', '{:.2f}'.format)


# Imputación de nulos usando métodos avanzados estadísticos
# -----------------------------------------------------------------------
from sklearn.impute import SimpleImputer
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer
from sklearn.impute import KNNImputer
from sklearn.ensemble import RandomForestRegressor

import warnings
warnings.filterwarnings("ignore")
pd.set_option('display.max_columns',100)


df = pd.read_csv("../../datos/02_api_rent_limpieza_cols.csv")
df.sample()

# Gestión de Valores erróneos
Vamos a revisarlos en:
- Precio
- pricebyarea
- size
- rooms
- bathrooms


df["province"].unique()

### Modelo 7 Eliminar casas ajenas a "Madrid" 

df = df[df["province"] == "Madrid"]
df.reset_index(drop=True,inplace=True)
df.shape

df["price"].unique()

No se ven anómalos

### size

df["size"].nlargest(5)

df["size"].nsmallest(5)

No parece erróneos

### rooms

df["rooms"].unique()

0 es que son estudios

### bathrooms

df["bathrooms"].unique()

No se ven valores erróneos

# Corregir "floor"
- Vamos a pasar todo a texto para que sea una categórica

df["floor"].unique()

### Generamos un diccionario para aplicar los cambios

diccionario_pisos = {
    "1": "primero",
    "2": "segundo",
    "3": "tercero",
    "4": "cuarto",
    "5": "quinto",
    "6": "sexto",
    "7": "septimo",
    "8": "octavo",
    "14": "decimo cuarto",
    "bj": "bajo",
    "en": "entreplanta",
    "ss": "sotano",
    "st": "sotano",
}

df["floor"] = df["floor"].map(diccionario_pisos)
df["floor"].unique()


### Modelo 13 Recategorizar pisos

bajos = ["primero","sotano","bajo","entreplanta"]
altos = ["segundo","tercero","cuarto","quinto","sexto","septimo","octavo","decimo cuarto"]

df.loc[df["floor"].isin(bajos), "floor"] = "bajos"
df.loc[df["floor"].isin(altos), "floor"] = "altos"

df["floor"].value_counts()

# Corregir Bathrooms
- Vamos a pasar todo a texto para que sea una categórica

df["bathrooms"].unique()

df.loc[df["bathrooms"] == 1, "bathrooms"] = "1 aseo"
df.loc[df["bathrooms"] == 2, "bathrooms"] = "2 aseos"
df.loc[df["bathrooms"] == 3, "bathrooms"] = "3 aseos"

df["bathrooms"].unique()

# Corregir rooms
- Vamos a pasar todo a texto para que sea una categórica

df["rooms"].unique()

df.loc[df["rooms"] == 0, "rooms"] = "sin habitaciones"
df.loc[df["rooms"] == 1, "rooms"] = "1 habitacion"
df.loc[df["rooms"] == 2, "rooms"] = "2 habitaciones"
df.loc[df["rooms"] == 3, "rooms"] = "3 habitaciones"
df.loc[df["rooms"] == 4, "rooms"] = "4 habitaciones"
df.loc[df["rooms"] == 6, "rooms"] = "6 habitaciones"

df["rooms"].unique()

# Reducción / Simplificación de columnas
- Vamos a quitarnos todo lo que:
    - Especifique demasiado para el precio
- Vamos a simplificar todo lo que:
    - Especifique demasiado, pero al generalizarlo, aporte valor

df.sample()

### Dropeamos
- address: No se puede generalizar

df.drop(columns="address",inplace=True)

### Generalizamos
- distance: Lo haremos por rangos de distancia al centro

df["distance"].nlargest()

df["distance"].nsmallest()

df["distancia_centro"] = "unknown"
df.loc[df["distance"].between(0,1000,inclusive="left"), "distancia_centro"] = "Menos de 1 km"
df.loc[df["distance"].between(1000,5000,inclusive="left"), "distancia_centro"] = "Entre 1 y 5 km"
df.loc[df["distance"].between(5000,10000,inclusive="left"), "distancia_centro"] =  "Entre 5 y 10 km"
df.loc[df["distance"].between(10000,20000,inclusive="left"), "distancia_centro"] = "Entre 10 y 20 km"
df.loc[df["distance"].between(20000,30000,inclusive="left"), "distancia_centro"] = "Entre 20 y 30 km"
df.loc[df["distance"].between(30000,40000,inclusive="left"), "distancia_centro"] = "Entre 30 y 40 km"
df.loc[df["distance"].between(40000,50000,inclusive="both"), "distancia_centro"] = "Entre 40 y 50 km"
df.loc[df["distance"] > 50000, "distancia_centro"] = "Mas de 50 km"

df["distancia_centro"].value_counts()

# Dropeamos distance

df.drop(columns="distance",inplace=True)

# Gestionar nulos variables categóricas
- Las booleanas las volvemos categóricas
- Los nans los pondremos en desconocido

df.info()

### propertyType

df["propertyType"].isnull().sum()

### status

df["status"].isnull().sum()

df["status"]= df["status"].fillna("desconocido")
df["status"].unique()

df["status"].isnull().sum()

### floor

df["floor"].isnull().sum()

df["floor"]= df["floor"].fillna("desconocido")
df["floor"].unique()

df["floor"].isnull().sum()

### neighborhood (Dropear Columna)
- Demasiados nulos, no tiene sentido

df["neighborhood"].isnull().sum()

df.drop(columns="neighborhood",inplace=True)

### district
- Tal vez especifique demasiado

df["district"].isnull().sum()

df["district"]= df["district"].fillna("desconocido")
df["district"].isnull().sum()

### province

df["province"].isnull().sum()

### municipality (Cambiado en Modelo 12)
- Vamos a reducir las opciones de la categoría
- Vamos a dividir por zonas en vez de municipios
    - Clase Alta
    - Clase Media-Alta
    - Clase Media
    - Clase Obrera

df["municipality"].isnull().sum()

clasificacion_municipios = {
    'Numancia de la Sagra': 'Clase Obrera',
    'Madrid': 'Clase Media-Alta',  # Varía según la zona
    'San Sebastián de los Reyes': 'Clase Alta',
    'Villamanrique de Tajo': 'Clase Obrera',
    'Recas': 'Clase Obrera',
    'Cedillo del Condado': 'Clase Obrera',
    'Rascafría': 'Clase Media',
    'Manzanares el Real': 'Clase Media',
    'Miraflores de la Sierra': 'Clase Media-Alta',
    'El Viso de San Juan': 'Clase Obrera',
    'Galapagar': 'Clase Media-Alta',
    'Arganda': 'Clase Media',
    'San Lorenzo de el Escorial': 'Clase Media-Alta',
    'Camarena': 'Clase Obrera',
    'Aldea del Fresno': 'Clase Obrera',
    'Aranjuez': 'Clase Media',
    'Villanueva del Pardillo': 'Clase Media-Alta',
    'Azuqueca de Henares': 'Clase Media',
    'El Espinar': 'Clase Media',
    'Las Rozas de Madrid': 'Clase Alta',
    'Guadalajara': 'Clase Media',
    'Illescas': 'Clase Obrera',
    'Navalcarnero': 'Clase Media',
    'Seseña': 'Clase Obrera',
    'Casarrubios del Monte': 'Clase Obrera',
    'Alcalá de Henares': 'Clase Media',
    'El Escorial': 'Clase Media-Alta',
    'Calypo Fado': 'Clase Obrera',
    'Leganés': 'Clase Media',
    'Coslada': 'Clase Media',
    'Torrejón de Ardoz': 'Clase Media',
    'Marchamalo': 'Clase Obrera',
    'Camarma de Esteruelas': 'Clase Media',
    'Alcorcón': 'Clase Media',
    'Pinto': 'Clase Media',
    'Pozo de Guadalajara': 'Clase Obrera',
    'Valdemoro': 'Clase Media',
    'Collado Villalba': 'Clase Media',
    'Getafe': 'Clase Media',
    'Paracuellos de Jarama': 'Clase Media-Alta',
    'El Molar': 'Clase Media',
    'Parla': 'Clase Obrera',
    'Tres Cantos': 'Clase Alta',
    'Yuncos': 'Clase Obrera',
    'Esquivias': 'Clase Obrera',
    'Quijorna': 'Clase Media',
    'Valdemorillo': 'Clase Media-Alta',
    'Yuncler': 'Clase Obrera',
    'Pedrezuela': 'Clase Media',
    'Daganzo de Arriba': 'Clase Media',
    'Yeles': 'Clase Obrera',
    'Guadarrama': 'Clase Media-Alta',
    'Ocaña': 'Clase Obrera',
    'Cobeña': 'Clase Media',
    'El Álamo': 'Clase Media',
    'Algete': 'Clase Media',
    'El Casar': 'Clase Obrera',
    'Rivas-Vaciamadrid': 'Clase Media-Alta',
    'Los Santos de la Humosa': 'Clase Media',
    'San Fernando de Henares': 'Clase Media',
    'Fuenlabrada': 'Clase Media',
    'Fuensalida': 'Clase Obrera',
    'Mataelpino': 'Clase Media',
    'Villa del Prado': 'Clase Media',
    'Mocejón': 'Clase Obrera',
    'Los Molinos': 'Clase Media',
    'Colmenar Viejo': 'Clase Media-Alta',
    'Móstoles': 'Clase Media',
    'Borox': 'Clase Obrera',
    'Navalafuente': 'Clase Media',
    'Meco': 'Clase Media',
    'Robledo de Chavela': 'Clase Media',
    'Campo Real': 'Clase Media',
    'Villaviciosa de Odón': 'Clase Alta',
    'San Ildefonso o la Granja': 'Clase Media',
    'Alameda de la Sagra': 'Clase Obrera',
    'Cabañas de la Sagra': 'Clase Obrera',
    'Las Navas del Marqués': 'Clase Media',
    'Villaseca de la Sagra': 'Clase Obrera',
    'Pozuelo de Alarcón': 'Clase Alta',
    'Yebes': 'Clase Obrera',
    'Bustarviejo': 'Clase Media',
    'Collado Mediano': 'Clase Media',
    'Chinchón': 'Clase Media',
    'Valmojado': 'Clase Obrera',
    'Alovera': 'Clase Media',
    'Colmenarejo': 'Clase Media-Alta',
    'Lominchar': 'Clase Obrera',
    'Loeches': 'Clase Media',
    'Sevilla la Nueva': 'Clase Media',
    'Serranillos del Valle': 'Clase Media',
    'Santa Cruz del Retamar': 'Clase Obrera',
    'Las Ventas de Retamosa': 'Clase Obrera',
    'Torrelaguna': 'Clase Media',
    'Villalbilla': 'Clase Media',
    'Alcobendas': 'Clase Alta',
}

# Asignar categorías al DataFrame
df["tipo_municipio"] = df["municipality"].map(clasificacion_municipios)

df["municipality"] = df["tipo_municipio"]
df.drop(columns="tipo_municipio", inplace=True)

df["municipality"].value_counts()

###  exterior
- Cambiamos de booleano a categorica

df["exterior"].isnull().sum()

df["exterior"].unique()

df.loc[df["exterior"] == True, "exterior"] = "vista exterior"
df.loc[df["exterior"] == False, "exterior"] = "vista interior"
df["exterior"].unique()

### hasLift 
- Cambiamos de booleano a categorica

df["hasLift"].isnull().sum()

df["hasLift"].unique()

df.loc[df["hasLift"] == True, "hasLift"] = "tiene ascensor"
df.loc[df["hasLift"] == False, "hasLift"] = "no tiene ascensor"
df["hasLift"] = df["hasLift"].fillna("desconocido")
df["hasLift"].unique()

df["hasLift"].isnull().sum()

### hasPlan
- Cambiamos de booleano a categorica


df["hasPlan"].isnull().sum()

df.loc[df["hasPlan"] == True, "hasPlan"] = "tiene planos"
df.loc[df["hasPlan"] == False, "hasPlan"] = "no tiene planos"
df["hasPlan"] = df["hasPlan"].fillna("desconocido")
df["hasPlan"].unique()

df["hasPlan"].isnull().sum()

### has3DTour
- Cambiamos de booleano a categorica


df["has3DTour"].isnull().sum()

df.loc[df["has3DTour"] == True, "has3DTour"] = "tiene tour 3D"
df.loc[df["has3DTour"] == False, "has3DTour"] = "no tiene tour 3D"
df["has3DTour"] = df["has3DTour"].fillna("desconocido")
df["has3DTour"].unique()

df["has3DTour"].isnull().sum()

### has360
- Cambiamos de booleano a categorica

df["has360"].isnull().sum()

c = df["has360"].isnull() == True
df[c]


#### Habrá que ver si estos datos merece la pena tenerlos, hay muchas cosas desconocidas

df.loc[df["has360"] == True, "has360"] = "tiene fotos 360"
df.loc[df["has360"] == False, "has360"] = "no tiene fotos 360"
df["has360"] = df["has360"].fillna("desconocido")
df["has360"].unique()

df["has360"].isnull().sum()

### distancia_centro

df["distancia_centro"].isnull().sum()

# Gestión nulos numéricas
- Sin tocar la VR (Precio).
- Rellenaremos con IterativeImputer y RandomForest
    - Las relaciones son complejas

df.isnull().sum()

# Modelo 7 eliminar province

df.drop(columns="province",inplace=True)

# Modelo 8 Eliminar district y has3DTour

df.drop(columns=["district","has3DTour"],inplace=True)

# Modelo 9 Eliminar status

df["status"].value_counts()

df.drop(columns="status",inplace=True)

# Modelo 11 Eliminar distancia_centro


df["distancia_centro"].value_counts()

#df.drop(columns="distancia_centro",inplace=True)

# Modelo 13 Eliminar has360 y hasPlan
- Considero que el usuario no tiene porque querer saber esto


df["has360"].value_counts()

df["hasPlan"].value_counts()

df.drop(columns=["has360","hasPlan"],inplace=True)

df["propertyType"].value_counts()

c1 = df["propertyType"] != "flat"
c2 = df["propertyType"] != "studio"

# Modelo 15 quitar "exterior"

df.drop(columns="exterior",inplace=True)

# Modelo 16 
- Usando property, se rellenan desconocidos de:
    - hasLift
    - floor



df["propertyType"].value_counts()

### que hago
- Todo lo que sea:
    - duplex
    - chalet
    - countryhouse
- No tienen ascensor y no tienen "floor" que se pondrá como "bajos"

- Los penthouse son áticos de lujo por lo que:
    - floor = altos
    - hasLift = si, porque es de lujo

tipos = ["duplex","chalet","countryHouse"]
df.loc[df["propertyType"].isin(tipos),"hasLift"] = "no tiene ascensor"
df.loc[df["propertyType"].isin(tipos),"floor"] = "bajos"
df.loc[df["propertyType"] == "penthouse","hasLift"] = "tiene ascensor"
df.loc[df["propertyType"] == "penthouse","floor"] = "altos"

df["floor"].value_counts()

df["hasLift"].value_counts()

### Eliminar propertyType y distancia_centro (Modelo 16)

df.drop(columns=["propertyType","distancia_centro"],inplace=True)

# Guardamos

df.to_csv("../../datos/03_api_rent_sin_nulos.csv",index=False)
